+++ UPM/asa_files/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 [See release notes](https://github.com/microsoft/MixedReality-WorldLockingTools-Unity/releases)
 
+## 1.4.1 - Workaround for OpenXR session tracking state issue.
+
 ## 1.4.0 - Initial release